[MinhDN] - Connfirmation message for deleting a keyword

+++ web/Keyword_JSP.jsp
@@ -248,7 +248,7 @@
                                                         <td><%= i + 1%></td>
                                                         <td class="keywords"><%= keywordDTO.getKeyword()%></td>
                                                         <td><%= keywordDTO.getUserId()%></td>
-                                                        <td><a href="MainController?btnAction=DeleteKeyword&id=<%= keywordDTO.getId()%>"><button><i class="pe-7s-trash" style="width: 20px; height: 20px"></i></button></a></td>
+                                                        <td><a href="MainController?btnAction=DeleteKeyword&id=<%= keywordDTO.getId()%>" onclick="return confirm('Are you sure you want to delete this item?');"><button><i class="pe-7s-trash" style="width: 20px; height: 20px"></i></button></a></td>
                                                     </tr>
                                                     <% } %>
                                                     <% }%>
